requirement.txt
langchain==1.0.0
langchain-groq==0.1.0
python-dotenv==1.0.0
streamlit==1.26.0
